Remove some `px` font-size overrides on topic metadata and buttons
After T314714, they are no longer needed for Vector, MonoBook etc. –
the elements will just use the normal body font size.

This changes the font size slightly on some skins (MonoBook, Timeless),
but I think that's okay.

We still want to customize the font-size for mobile (Minerva), and
we can do that with `em` now.

Bug: T325104
Change-Id: I3fab8588b1b0e3a43aac4c32a3610f64fb3d3fb3