Add `{{GENDER:|}}` to some localisation messages
Although it's not necessary to use `{{GENDER:|}}` (with empty first
parameter – referring to the user viewing the page) in the English
message to enable its use in translations, translators keep asking us
about it, so let's make it clear.

Add the markup to messages that already use it in at least one
translation and those phrased very similarly to one of them.
See task for details.

Bug: T332164
Change-Id: I27ca34c41493921d4c7fda8f92f40c3495af2804