build: Rewrite exclude rule Generic.CodeAnalysis.AssignmentInCondition
The <exclude name=".."/> is managed by libup and when a rule passed,
it get removed from the list. Failing rules on update are added.

Setting the severity for the rule to none avoid that this is changed by
libup and it is more visible, that the disable is wanted and there is
nothing to "fix" as usually assumend for sniffs in the exclude list.

Follow-Up: I8adfdce01ea96f4b62dabd3dea130f9593c7e5ac
Change-Id: If42dade02c6ceed1e00f5a3dc846e1578b960ca6