Always set a value for isEmptyTalkPage when transforming html
For interface messages, like the talk page header, tranformHtml isn't
run so no value is set while parsing it.  The double negative in
I5c1877f7f9eb73f88a33e001ca3c2f3d06bb90e4 failed to account for that and
was reverted in I23a3f937201d93c7c7645a09c4fccfcf1c14008a.

By always setting a value, we'll overwrite a value that may have been
set when Parsoid calls the legacy parser to parse extension content for
which it doesn't have a native extension, which then gets merged into
the metadata for the main parse.  However, setting with a different
value is deprecated and will eventually throw, but we already need to
find a solution for T372592 so what's one more.

Bug: T371125
Bug: T372592
Change-Id: Id73a1b5751cfc055e84188bcb19583c72b84032f