Notifications: Handle suppressed topic titles in notification
EchoPresentationModelSection::getTruncatedSectionTitle() has a note that
it should not be called if ::exists() returned false. Among the reason
the section may not exist includes suppressing the revision where the
topic title was added or edited.

Add placeholder message text "(topic removed)", to be used in place of
the unavailable topic title. This is analogous to "(user removed)" which
would be shown along if the username is also suppressed.

Bug: T378261
Change-Id: I6db0872b8724323a5e15eb6a262c2c3d0bdd630d