Add tests for empty state hooks
Assert that the relevant message appears, and that it only appears
once, under different conditions.

These tests are big ugly integration tests and they're pretty slow,
but I think it's worth it in this case, as we've had many issues in
the past where the message did not appear or appeared twice, and the
code is still fragile.

Bug: T395196
Change-Id: I5fd57f4345bf2cf1ef7d808de827fa088aa59142