Prevent default when clicking on permalinks
Previously we let the link's default navigation happen which,
if you were at the canonical URL for the talk page, would not
cause the page to reload. However if you had anything in your
query string, or are on the mobile domain, then the page reloads
to take you to the canonical version of the URL.

This isn't really necessary as users likely copying the URL by
just left clicking (in which case we write to the clipboard with
JS) or right clicking and choosing "Copy link address". In either
case they get the canonical permalink for the comment. They are
unlikely to click on the permalink and then expect to copy it
from the address bar. Even if a user did this, it isn't much
of a problem as a non-canonical URL will still work.

As a result we can just update the current location.hash, meaning
we never reload the page, and rely on our JS to write the
canonical URL to the clipboard.

Bug: T388346
Bug: T396138
Change-Id: Ic9084d039e94a2b8c15b6bbff23742bc1eff9e85