tests: skip tests when MobileFrontend is not loaded
On CI, we would like to remove the recursive injection of dependencies.
CommentFormatterTest uses MobileFrontend for an integration test, skip
the tests when MobileFrontend is not loaded.

The CI configuration adds MobileFrontend as a dependency of
DiscussionTools, it is thus remains available for patches to
that extension.

Bug: T411339
Change-Id: I09f57b78def4a757c80fe476b2ef7abb18b22c54
(cherry picked from commit c0b770909c19c056874155025ff0a71d090ed9b9)