Make scroll padding computation skin-agnostic
Vector no longer uses the class vector-feature-sticky-header-enabled.
Instead we can just use the computed value of scroll-padding-top.

Also increase the amount of padding above the reply tool
we attempt to scroll back into view so in the common case
we can see a bit of the comment being replied to.

Bug: T406120
Change-Id: I02a5222a0273c53573134469533c64312c6a40fa