Remove DiscussionToolsAutoTopicSubEditor config option
The 'discussiontoolsapi'-only mode was intended as a
transitional setting while we introduced to the community
to the concept of topic subscriptions.

After I9ef7dcd21f30 this will be set to the default
value of 'any' on all WMF wikis so keeping the other
mode around is an unnecessary maintenance burden.

Change-Id: I01751c034727df2d40ad328b8571c60c163fc30a